moved unit to metrics sub-package
